+++ CHANGELOG.md
@@ -8,6 +8,7 @@ See [Milestones](https://github.com/sinus-x/rubbergoddess/milestones) to get an 
 
 ### Added
 
+- Account cog: Manage bot's user account.
 - Points cog: MEE6-like points.
 - Seeking cog: Announce that you're seeking something/someone.